Move pkg ts/m3db to m3/storage
Previously m3/storage depended on ts/m3db for m3db.Options. This made it
impossible for m3db.Options to include any exposed types from
m3/storage, like ComputedNamespace.
