+++ src/query/functions/binary/and_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/m3db/m3/src/query/models"
 	"github.com/m3db/m3/src/query/parser"
 	"github.com/m3db/m3/src/query/test"
+	"github.com/m3db/m3/src/query/test/compare"
 	"github.com/m3db/m3/src/query/test/executor"
 	xtime "github.com/m3db/m3/src/x/time"
 
@@ -99,7 +100,7 @@ func TestAndWithSomeValues(t *testing.T) {
 	expected := values1
 	expected[0][1] = math.NaN()
 	expected[1][0] = math.NaN()
-	test.EqualsWithNans(t, expected, sink.Values)
+	compare.EqualsWithNans(t, expected, sink.Values)
 }
 
 var andTests = []struct {
@@ -233,7 +234,7 @@ func TestAnd(t *testing.T) {
 			}
 
 			require.NoError(t, err)
-			test.EqualsWithNans(t, tt.expected, sink.Values)
+			compare.EqualsWithNans(t, tt.expected, sink.Values)
 			meta := sink.Meta
 			assert.Equal(t, 0, meta.Tags.Len())
 			assert.True(t, meta.Bounds.Equals(bounds))
