+++ src/query/functions/binary/binary_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/m3db/m3/src/query/models"
 	"github.com/m3db/m3/src/query/parser"
 	"github.com/m3db/m3/src/query/test"
+	"github.com/m3db/m3/src/query/test/compare"
 	"github.com/m3db/m3/src/query/test/executor"
 	xtime "github.com/m3db/m3/src/x/time"
 
@@ -139,7 +140,7 @@ func TestScalars(t *testing.T) {
 				tt.expected, tt.expected,
 			}}
 
-			test.EqualsWithNans(t, expected, sink.Values)
+			compare.EqualsWithNans(t, expected, sink.Values)
 
 			assert.Equal(t, bounds, sink.Meta.Bounds)
 			assert.Equal(t, 0, sink.Meta.Tags.Len())
@@ -203,7 +204,7 @@ func TestScalarsReturnBoolFalse(t *testing.T) {
 				tt.expected, tt.expected,
 			}}
 
-			test.EqualsWithNans(t, expected, sink.Values)
+			compare.EqualsWithNans(t, expected, sink.Values)
 
 			assert.Equal(t, bounds, sink.Meta.Bounds)
 			assert.Equal(t, 0, sink.Meta.Tags.Len())
@@ -597,7 +598,7 @@ func TestSingleSeriesReturnBool(t *testing.T) {
 				require.NoError(t, err)
 			}
 
-			test.EqualsWithNans(t, tt.expectedBool, sink.Values)
+			compare.EqualsWithNans(t, tt.expectedBool, sink.Values)
 
 			assert.Equal(t, bounds, sink.Meta.Bounds)
 			assert.Equal(t, 0, sink.Meta.Tags.Len())
@@ -665,7 +666,7 @@ func TestSingleSeriesReturnValues(t *testing.T) {
 				require.NoError(t, err)
 			}
 
-			test.EqualsWithNans(t, tt.expected, sink.Values)
+			compare.EqualsWithNans(t, tt.expected, sink.Values)
 
 			assert.Equal(t, bounds, sink.Meta.Bounds)
 			assert.Equal(t, 0, sink.Meta.Tags.Len())
@@ -920,7 +921,7 @@ func TestBothSeries(t *testing.T) {
 				test.NewBlockFromValuesWithSeriesMeta(bounds, tt.rhsMeta, tt.rhs))
 			require.NoError(t, err)
 
-			test.EqualsWithNans(t, tt.expected, sink.Values)
+			compare.EqualsWithNans(t, tt.expected, sink.Values)
 
 			// Extract duped expected metas
 			expectedMeta := block.Metadata{
@@ -1003,7 +1004,7 @@ func TestBinaryFunctionWithDifferentNames(t *testing.T) {
 	err = node.Process(models.NoopQueryContext(), parser.NodeID(rune(1)), right)
 	require.NoError(t, err)
 
-	test.EqualsWithNans(t, expected, sink.Values)
+	compare.EqualsWithNans(t, expected, sink.Values)
 
 	exResultMeta := block.ResultMetadata{
 		LocalOnly:  false,
@@ -1095,7 +1096,7 @@ func TestOneToOneMatcher(t *testing.T) {
 	err = node.Process(models.NoopQueryContext(), parser.NodeID(rune(1)), right)
 	require.NoError(t, err)
 
-	test.EqualsWithNans(t, expected, sink.Values)
+	compare.EqualsWithNans(t, expected, sink.Values)
 
 	expectedMetas := []block.SeriesMeta{
 		{
