+++ src/query/storage/m3/encoded_step_iterator.go
@@ -18,11 +18,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package m3db
+package m3
 
 import (
 	"github.com/m3db/m3/src/query/block"
-	"github.com/m3db/m3/src/query/ts/m3db/consolidators"
+	"github.com/m3db/m3/src/query/storage/m3/consolidators"
 )
 
 type encodedStepIter struct {
