+++ src/query/test/compare/comparison.go
@@ -18,7 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package test
+// Package compare provides comparison functions for testing.
+package compare
 
 import (
 	"bytes"
