+++ src/query/test/m3/test_storage.go
@@ -28,7 +28,6 @@ import (
 	"github.com/m3db/m3/src/query/models"
 	"github.com/m3db/m3/src/query/storage"
 	"github.com/m3db/m3/src/query/storage/m3"
-	"github.com/m3db/m3/src/query/ts/m3db"
 	"github.com/m3db/m3/src/x/ident"
 	"github.com/m3db/m3/src/x/instrument"
 	"github.com/m3db/m3/src/x/sync"
@@ -67,7 +66,7 @@ func NewStorageAndSession(
 	require.NoError(t, err)
 	writePool.Init()
 	tagOptions := models.NewTagOptions().SetMetricName([]byte("name"))
-	opts := m3db.NewOptions().
+	opts := m3.NewOptions().
 		SetWriteWorkerPool(writePool).
 		SetTagOptions(tagOptions).
 		SetLookbackDuration(defaultLookbackDuration)
@@ -101,7 +100,7 @@ func NewStorageAndSessionWithAggregatedNamespaces(
 	require.NoError(t, err)
 	writePool.Init()
 	tagOptions := models.NewTagOptions().SetMetricName([]byte("name"))
-	opts := m3db.NewOptions().
+	opts := m3.NewOptions().
 		SetWriteWorkerPool(writePool).
 		SetTagOptions(tagOptions).
 		SetLookbackDuration(defaultLookbackDuration)
